Fix from Develop (Character - Champion) + Projectiles Tags
