+++ Assets/Scripts/Champions/Windiator.meta
@@ -1,10 +1,9 @@
 fileFormatVersion: 2
-guid: 3f76ad1263ede2746ae86272b177e6bb
+guid: c0a8bf592f6d79542b6af08ec1225848
 folderAsset: yes
-timeCreated: 1515329545
+timeCreated: 1510328916
 licenseType: Free
 DefaultImporter:
-  externalObjects: {}
   userData: 
   assetBundleName: 
   assetBundleVariant:
