+++ Assets/Scripts/Champions/Windiator/AutoAttack.meta
@@ -1,10 +1,9 @@
 fileFormatVersion: 2
-guid: 3f76ad1263ede2746ae86272b177e6bb
+guid: b08dc89f0cd22c34a804e0316504d899
 folderAsset: yes
-timeCreated: 1515329545
+timeCreated: 1510334053
 licenseType: Free
 DefaultImporter:
-  externalObjects: {}
   userData: 
   assetBundleName: 
   assetBundleVariant:
