+++ Assets/Scripts/Champions/Windiator/Passive.meta
@@ -1,10 +1,9 @@
 fileFormatVersion: 2
-guid: 3f76ad1263ede2746ae86272b177e6bb
+guid: 2f7088c23fa29ca43abe04ee015cde50
 folderAsset: yes
-timeCreated: 1515329545
+timeCreated: 1510334145
 licenseType: Free
 DefaultImporter:
-  externalObjects: {}
   userData: 
   assetBundleName: 
   assetBundleVariant:
