+++ Assets/Scripts/Champions/Windiator/ScriptsOnPrefabs.meta
@@ -1,10 +1,9 @@
 fileFormatVersion: 2
-guid: 3f76ad1263ede2746ae86272b177e6bb
+guid: 5b949112da1c61a4582b107443c94ae7
 folderAsset: yes
-timeCreated: 1515329545
+timeCreated: 1510334174
 licenseType: Free
 DefaultImporter:
-  externalObjects: {}
   userData: 
   assetBundleName: 
   assetBundleVariant:
