+++ Assets/Scripts/Champions/Windiator/ScriptsOnPrefabs/StormWallBehaviour.cs.meta
