+++ Assets/Scripts/Champions/Windiator/ScriptsOnPrefabs/Trident.cs
