+++ Assets/Scripts/Champions/Windiator/ScriptsOnPrefabs/WindSpiritFavorStatus.cs.meta
