+++ Assets/Scripts/Champions/Windiator/Spells.meta
@@ -1,10 +1,9 @@
 fileFormatVersion: 2
-guid: 3f76ad1263ede2746ae86272b177e6bb
+guid: 9ba154cb6cb424340a8085afadbce026
 folderAsset: yes
-timeCreated: 1515329545
+timeCreated: 1510334205
 licenseType: Free
 DefaultImporter:
-  externalObjects: {}
   userData: 
   assetBundleName: 
   assetBundleVariant:
