+++ Assets/Scripts/Classes/BASE_CHARS.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3f76ad1263ede2746ae86272b177e6bb
+guid: 878bcbb6da7065c458c01092f554333e
 folderAsset: yes
 timeCreated: 1515329545
 licenseType: Free
