+++ Assets/Scripts/Classes/Windiator.meta
@@ -1,9 +1,10 @@
 fileFormatVersion: 2
-guid: c0a8bf592f6d79542b6af08ec1225848
+guid: e1e670c439904e94d9e4c9fabf997d6e
 folderAsset: yes
-timeCreated: 1510328916
+timeCreated: 1515329545
 licenseType: Free
 DefaultImporter:
+  externalObjects: {}
   userData: 
   assetBundleName: 
   assetBundleVariant:
