+++ Assets/Scripts/Classes/Windiator/Passive.meta
@@ -1,9 +1,10 @@
 fileFormatVersion: 2
-guid: 2f7088c23fa29ca43abe04ee015cde50
+guid: a624ecd9a33b8b84fbb027dfcd58a0cb
 folderAsset: yes
-timeCreated: 1510334145
+timeCreated: 1515329545
 licenseType: Free
 DefaultImporter:
+  externalObjects: {}
   userData: 
   assetBundleName: 
   assetBundleVariant:
