+++ Assets/Scripts/Classes/Windiator/ScriptsOnPrefabs.meta
@@ -1,9 +1,10 @@
 fileFormatVersion: 2
-guid: 5b949112da1c61a4582b107443c94ae7
+guid: 5ff8826b199281e4fb5d09d2c8332115
 folderAsset: yes
-timeCreated: 1510334174
+timeCreated: 1515329545
 licenseType: Free
 DefaultImporter:
+  externalObjects: {}
   userData: 
   assetBundleName: 
   assetBundleVariant:
