+++ Assets/Scripts/Classes/Windiator/Spells.meta
@@ -1,9 +1,10 @@
 fileFormatVersion: 2
-guid: 9ba154cb6cb424340a8085afadbce026
+guid: c952a543f3ca784419ece00bc2327590
 folderAsset: yes
-timeCreated: 1510334205
+timeCreated: 1515329545
 licenseType: Free
 DefaultImporter:
+  externalObjects: {}
   userData: 
   assetBundleName: 
   assetBundleVariant:
